webpage_to_markdown/requirements.txt
playwright
openai
python-dotenv
Pillow